design/lasercut/make_3d/requirements.txt
ezdxf
solidpython
